+++ index.js
@@ -1,4 +1,4 @@
 require('rootpath')();
 require('@babel/register');
 
-module.exports = require('./server.jsx');
+module.exports = require('./server');