+++ src/app/utils/SearchModel.js
@@ -12,7 +12,7 @@ import exceptionalDomains from './ExceptionalDomains';
  * fetchResultLength(data)
  * The function gets the total search result number.
  *
- * @param {Object} item
+ * @param {Object} data
  * @return {Num}
  */
 const fetchResultLength = (data) => {
@@ -65,6 +65,7 @@ const extractSearchElements = (requestCombo) => {
 };
 
 const displayName = (name) => filterNames.find((obj) => obj.value === name).anchor;
+
 /**
  * fetchDisplayName(labelsArray, searchRequest)
  * The function returns the display name of the item.
@@ -148,7 +149,6 @@ const secureHttpsProtocol = (url, domains) => {
  *
  * @param {Object} item
  * @param {String} feature
- * @param {String} SearchRequest
  * @return {String}
  */
 const fetchItemFeature = (item, feature) => (item[feature] || '');
@@ -178,7 +178,7 @@ const fetchThumbnail = (item) => {
  * to call extractSearchElements(searchRequest) to get the present facet.
  *
  * @param {Object} item
- * @param {String} SearchRequest
+ * @param {String} searchRequest
  * @return {Object}
  */
 const fetchItem = (item, searchRequest) => {
@@ -212,7 +212,7 @@ const fetchItem = (item, searchRequest) => {
  * It returns an array with each item inside.
  *
  * @param {Object} data
- * @param {String} SearchRequest
+ * @param {String} searchRequest
  * @return {Array}
  */
 const fetchResultItems = (data, searchRequest = '') => {