+++ src/server/ApiRoutes/cacheUtil.js
@@ -13,31 +13,26 @@ const { redisHosts } = appConfig;
 const getKeyFromParams = (params) => params.map((x) => JSON.stringify(x)).join(',');
 
 /**
- checkForKeyInRedis(client)
- Takes a caching client (e.g. redis). Returns a function which takes a key and
- returns a promise that either resolves to the stored value for that key, or null
- if there is no stored value.
-
- @param {object} client
- @param {string} key
+ * checkForKeyInRedis(client)
+ * Takes a caching client (e.g. redis). Returns a function which takes a key and
+ * returns a promise that either resolves to the stored value for that key, or null
+ * if there is no stored value.
+ *
+ * @param {object} client
  */
-
 const checkForKeyInRedis = (client) => (key) => new Promise((resolve) => client
   .get(key, (err, cachedResponse) => resolve(cachedResponse)));
 
 /**
- getDataAndSetKeyInClient(dataFunction, client)
- Takes a dataFunction, which fetches new data (e.g. a function which calls axios)
- and a caching client (e.g. redis). Returns a function which takes some params
- and a key, fetches the corresponding data for those params, and caches that
- data under the given key
-
- @param {function} dataFunction
- @param {object} client
- @param {array} params
- @param {string} key
+ * getDataAndSetKeyInClient(dataFunction, client)
+ * Takes a dataFunction, which fetches new data (e.g. a function which calls axios)
+ * and a caching client (e.g. redis). Returns a function which takes some params
+ * and a key, fetches the corresponding data for those params, and caches that
+ * data under the given key
+ *
+ * @param {function} dataFunction
+ * @param {object} client
  */
-
 const getDataAndSetKeyInClient = (dataFunction, client) => (params, key) => dataFunction(...params)
   .then((response) => {
     const stringifiedResponse = JSON.stringify(response);
@@ -47,15 +42,14 @@ const getDataAndSetKeyInClient = (dataFunction, client) => (params, key) => data
   });
 
 /**
- useCachedOrGetData(dataFunction, client)
- Takes a dataFunction, which fetches new data (e.g. a function which calls axios)
- and a caching client (e.g. redis). Returns a function which takes params, checks
- to see if those params are cached, and then either returns the cached value or gets
- a new value for those params if none is cached
-
- @param {function} dataFunction
- @param {object} client
- @param {array} params
+ * useCachedOrGetData(dataFunction, client)
+ * Takes a dataFunction, which fetches new data (e.g. a function which calls axios)
+ * and a caching client (e.g. redis). Returns a function which takes params, checks
+ * to see if those params are cached, and then either returns the cached value or gets
+ * a new value for those params if none is cached
+ *
+ * @param {function} dataFunction
+ * @param {object} client
  */
 
 const useCachedOrGetData = (dataFunction, client) => (params) => {
@@ -99,21 +93,20 @@ const generateClient = (customClient = null, appEnv = null, region = 'us-east-1'
 };
 
 /**
- addCaching(dataFunction, useClient, customClient, appEnv, region)
- Returns a promise which will resolve to a function that wraps the given
- datafunction. The wrapper will call useCachedOrGetData with the given datafunction
- and to ensure the value is cached, and then return the value.
- addCaching will also generate the dataclient if a custom client is not given and
- useClient is not set to false. appEnv and region are passed to generateClient.
-
- @param {function} datafunction
- @param {boolean} useClient - Controls whether or not to use caching
-                              (in practice, set to !process.env.USE_CACHING)
- @param {object} customClient
- @param {string} appEnv
- @param {string} region
+ * addCaching(dataFunction, useClient, customClient, appEnv, region)
+ * Returns a promise which will resolve to a function that wraps the given
+ * datafunction. The wrapper will call useCachedOrGetData with the given datafunction
+ * and to ensure the value is cached, and then return the value.
+ * addCaching will also generate the dataclient if a custom client is not given and
+ * useClient is not set to false. appEnv and region are passed to generateClient.
+ *
+ * @param {function} dataFunction
+ * @param {boolean} useClient - Controls whether or not to use caching
+ *                              (in practice, set to !process.env.USE_CACHING)
+ * @param {?Object} customClient
+ * @param {string} appEnv
+ * @param {string} region
  */
-
 const addCaching = (dataFunction, useClient = true, customClient = null, appEnv, region = 'us-east-1') => {
   if (!useClient) {
     return Promise.resolve(dataFunction);